# internal/treasury/api/http/openapi.gen.go
// Package http provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.6.0 DO NOT EDIT.
package http

import (
	"context"
	"encoding/json"
	"fmt"
	"net/http"
	"time"

	"sumni-finance-backend/internal/common/shared"
	"sumni-finance-backend/internal/treasury/domain"

	"github.com/labstack/echo/v4"
	"github.com/oapi-codegen/runtime"
	strictecho "github.com/oapi-codegen/runtime/strictmiddleware/echo"
	"github.com/shopspring/decimal"
)

// Currency defines model for Currency.
type Currency = shared.Currency

// Decimal defines model for Decimal.
type Decimal = decimal.Decimal

// EntryType Journal entry type (DEBIT || CREDIT)
type EntryType = shared.EntryType

// ErrorDetails defines model for ErrorDetails.
type ErrorDetails struct {
	EntityId   *string `json:"entity_id,omitempty"`
	EntityType *string `json:"entity_type,omitempty"`
	ErrorSlug  string  `json:"error_slug"`
	Message    string  `json:"message"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Details Detailed error information
	Details []ErrorDetails `json:"details"`

	// Message Human-readable error message
	Message string `json:"message"`

	// Slug Error slug
	Slug string `json:"slug"`
}

// FundSourceBankMetadata defines model for FundSourceBankMetadata.
type FundSourceBankMetadata struct {
	AccountNumber string `json:"account_number"`
	AccountOwner  string `json:"account_owner"`
	BankCode      string `json:"bank_code"`
}

// FundSourceCashMetadata defines model for FundSourceCashMetadata.
type FundSourceCashMetadata struct {
	OwnerName string `json:"owner_name"`
}

// FundSourceMetadata defines model for FundSourceMetadata.
type FundSourceMetadata struct {
	BankMetadata *FundSourceBankMetadata `json:"bank_metadata,omitempty"`
	CashMetadata *FundSourceCashMetadata `json:"cash_metadata,omitempty"`
}

// FundSourceType Fund source type (CASH || BANK)
type FundSourceType = domain.FundSourceType

// FundSourceUUID UUID of a fund source
type FundSourceUUID = domain.FundSourceUUID

// JournalEntryItem defines model for JournalEntryItem.
type JournalEntryItem struct {
	Amount      Decimal `json:"amount"`
	Description *string `json:"description,omitempty"`

	// EntryType Journal entry type (DEBIT || CREDIT)
	EntryType       EntryType `json:"entry_type"`
	TransactionDate time.Time `json:"transaction_date"`
	TransactionNo   *string   `json:"transaction_no,omitempty"`
}

// JournalEntryResponse defines model for JournalEntryResponse.
type JournalEntryResponse struct {
	Amount        Decimal   `json:"amount"`
	BalanceAfter  Decimal   `json:"balance_after"`
	BalanceBefore Decimal   `json:"balance_before"`
	CreatedAt     time.Time `json:"created_at"`
	CreatedBy     string    `json:"created_by"`
	Description   *string   `json:"description,omitempty"`

	// EntryType Journal entry type (DEBIT || CREDIT)
	EntryType EntryType `json:"entry_type"`

	// FundSourceUuid UUID of a fund source
	FundSourceUuid FundSourceUUID `json:"fund_source_uuid"`

	// JournalEntryUuid UUID of a journal entry
	JournalEntryUuid JournalEntryUUID `json:"journal_entry_uuid"`

	// ReverseEntryUuid UUID of a journal entry
	ReverseEntryUuid *JournalEntryUUID `json:"reverse_entry_uuid,omitempty"`

	// Status Journal entry status (RECORDED || POSTED || VOIDED)
	Status          JournalEntryStatus `json:"status"`
	TransactionDate time.Time          `json:"transaction_date"`
	TransactionNo   *string            `json:"transaction_no,omitempty"`
	UpdatedAt       *time.Time         `json:"updated_at,omitempty"`
	UpdatedBy       *string            `json:"updated_by,omitempty"`
	VoidedAt        *time.Time         `json:"voided_at,omitempty"`
	VoidedBy        *string            `json:"voided_by,omitempty"`
	VoidedReason    *string            `json:"voided_reason,omitempty"`
}

// JournalEntryStatus Journal entry status (RECORDED || POSTED || VOIDED)
type JournalEntryStatus = domain.JournalEntryStatus

// JournalEntryUUID UUID of a journal entry
type JournalEntryUUID = domain.JournalEntryUUID

// RecordJournalEntriesRequest defines model for RecordJournalEntriesRequest.
type RecordJournalEntriesRequest = []JournalEntryItem

// RecordJournalEntriesResponse defines model for RecordJournalEntriesResponse.
type RecordJournalEntriesResponse = map[string]interface{}

// RegisterFundSourceRequest defines model for RegisterFundSourceRequest.
type RegisterFundSourceRequest struct {
	Currency    Currency           `json:"currency"`
	InitBalance Decimal            `json:"init_balance"`
	Metadata    FundSourceMetadata `json:"metadata"`
	Name        string             `json:"name"`

	// SourceType Fund source type (CASH || BANK)
	SourceType FundSourceType `json:"source_type"`
}

// RegisterFundSourceResponse defines model for RegisterFundSourceResponse.
type RegisterFundSourceResponse struct {
	// FundSourceUuid UUID of a fund source
	FundSourceUuid FundSourceUUID `json:"fund_source_uuid"`
}

// VoidJournalEntryRequest defines model for VoidJournalEntryRequest.
type VoidJournalEntryRequest struct {
	VoidReason string `json:"void_reason"`
}

// VoidJournalEntryResponse defines model for VoidJournalEntryResponse.
type VoidJournalEntryResponse struct {
	ReverseJournalEntry JournalEntryResponse `json:"reverse_journal_entry"`
}

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// InternalServerError defines model for InternalServerError.
type InternalServerError = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// RegisterFundSourceJSONRequestBody defines body for RegisterFundSource for application/json ContentType.
type RegisterFundSourceJSONRequestBody = RegisterFundSourceRequest

// RecordJournalEntriesJSONRequestBody defines body for RecordJournalEntries for application/json ContentType.
type RecordJournalEntriesJSONRequestBody = RecordJournalEntriesRequest

// VoidJournalEntryJSONRequestBody defines body for VoidJournalEntry for application/json ContentType.
type VoidJournalEntryJSONRequestBody = VoidJournalEntryRequest

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Register a new fund source
	// (POST /v1/treasury/fund-sources)
	RegisterFundSource(ctx echo.Context) error
	// Record Journal Entries (the fund source's balance will be affected)
	// (POST /v1/treasury/fund-sources/{fund_source_uuid}/journal-entries)
	RecordJournalEntries(ctx echo.Context, fundSourceUuid FundSourceUUID) error
	// Void a journal entry by creating a reverse journal entry
	// (POST /v1/treasury/fund-sources/{fund_source_uuid}/journal-entries/{journal_entry_uuid}/void)
	VoidJournalEntry(ctx echo.Context, fundSourceUuid FundSourceUUID, journalEntryUuid JournalEntryUUID) error
}

// ServerInterfaceWrapper converts echo contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler ServerInterface
}

// RegisterFundSource converts echo context to params.
func (w *ServerInterfaceWrapper) RegisterFundSource(ctx echo.Context) error {
	var err error

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.RegisterFundSource(ctx)
	return err
}

// RecordJournalEntries converts echo context to params.
func (w *ServerInterfaceWrapper) RecordJournalEntries(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "fund_source_uuid" -------------
	var fundSourceUuid FundSourceUUID

	err = runtime.BindStyledParameterWithOptions("simple", "fund_source_uuid", ctx.Param("fund_source_uuid"), &fundSourceUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter fund_source_uuid: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.RecordJournalEntries(ctx, fundSourceUuid)
	return err
}

// VoidJournalEntry converts echo context to params.
func (w *ServerInterfaceWrapper) VoidJournalEntry(ctx echo.Context) error {
	var err error
	// ------------- Path parameter "fund_source_uuid" -------------
	var fundSourceUuid FundSourceUUID

	err = runtime.BindStyledParameterWithOptions("simple", "fund_source_uuid", ctx.Param("fund_source_uuid"), &fundSourceUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter fund_source_uuid: %s", err))
	}

	// ------------- Path parameter "journal_entry_uuid" -------------
	var journalEntryUuid JournalEntryUUID

	err = runtime.BindStyledParameterWithOptions("simple", "journal_entry_uuid", ctx.Param("journal_entry_uuid"), &journalEntryUuid, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true, Type: "string", Format: "uuid"})
	if err != nil {
		return echo.NewHTTPError(http.StatusBadRequest, fmt.Sprintf("Invalid format for parameter journal_entry_uuid: %s", err))
	}

	// Invoke the callback with all the unmarshaled arguments
	err = w.Handler.VoidJournalEntry(ctx, fundSourceUuid, journalEntryUuid)
	return err
}

// This is a simple interface which specifies echo.Route addition functions which
// are present on both echo.Echo and echo.Group, since we want to allow using
// either of them for path registration
type EchoRouter interface {
	CONNECT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	DELETE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	GET(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	HEAD(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	OPTIONS(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PATCH(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	POST(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	PUT(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
	TRACE(path string, h echo.HandlerFunc, m ...echo.MiddlewareFunc) *echo.Route
}

// RegisterHandlers adds each server route to the EchoRouter.
func RegisterHandlers(router EchoRouter, si ServerInterface) {
	RegisterHandlersWithBaseURL(router, si, "")
}

// Registers handlers, and prepends BaseURL to the paths, so that the paths
// can be served under a prefix.
func RegisterHandlersWithBaseURL(router EchoRouter, si ServerInterface, baseURL string) {

	wrapper := ServerInterfaceWrapper{
		Handler: si,
	}

	router.POST(baseURL+"/v1/treasury/fund-sources", wrapper.RegisterFundSource)
	router.POST(baseURL+"/v1/treasury/fund-sources/:fund_source_uuid/journal-entries", wrapper.RecordJournalEntries)
	router.POST(baseURL+"/v1/treasury/fund-sources/:fund_source_uuid/journal-entries/:journal_entry_uuid/void", wrapper.VoidJournalEntry)

}

type BadRequestJSONResponse ErrorResponse

type InternalServerErrorJSONResponse ErrorResponse

type NotFoundJSONResponse ErrorResponse

type RegisterFundSourceRequestObject struct {
	Body *RegisterFundSourceJSONRequestBody
}

type RegisterFundSourceResponseObject interface {
	VisitRegisterFundSourceResponse(w http.ResponseWriter) error
}

type RegisterFundSource201JSONResponse RegisterFundSourceResponse

func (response RegisterFundSource201JSONResponse) VisitRegisterFundSourceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type RegisterFundSource400JSONResponse struct{ BadRequestJSONResponse }

func (response RegisterFundSource400JSONResponse) VisitRegisterFundSourceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type RegisterFundSource409JSONResponse ErrorResponse

func (response RegisterFundSource409JSONResponse) VisitRegisterFundSourceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)

	return json.NewEncoder(w).Encode(response)
}

type RegisterFundSource500JSONResponse struct {
	InternalServerErrorJSONResponse
}

func (response RegisterFundSource500JSONResponse) VisitRegisterFundSourceResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type RecordJournalEntriesRequestObject struct {
	FundSourceUuid FundSourceUUID `json:"fund_source_uuid"`
	Body           *RecordJournalEntriesJSONRequestBody
}

type RecordJournalEntriesResponseObject interface {
	VisitRecordJournalEntriesResponse(w http.ResponseWriter) error
}

type RecordJournalEntries201JSONResponse RecordJournalEntriesResponse

func (response RecordJournalEntries201JSONResponse) VisitRecordJournalEntriesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(201)

	return json.NewEncoder(w).Encode(response)
}

type RecordJournalEntries400JSONResponse struct{ BadRequestJSONResponse }

func (response RecordJournalEntries400JSONResponse) VisitRecordJournalEntriesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type RecordJournalEntries404JSONResponse struct{ NotFoundJSONResponse }

func (response RecordJournalEntries404JSONResponse) VisitRecordJournalEntriesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type RecordJournalEntries409JSONResponse ErrorResponse

func (response RecordJournalEntries409JSONResponse) VisitRecordJournalEntriesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(409)

	return json.NewEncoder(w).Encode(response)
}

type RecordJournalEntries500JSONResponse struct {
	InternalServerErrorJSONResponse
}

func (response RecordJournalEntries500JSONResponse) VisitRecordJournalEntriesResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

type VoidJournalEntryRequestObject struct {
	FundSourceUuid   FundSourceUUID   `json:"fund_source_uuid"`
	JournalEntryUuid JournalEntryUUID `json:"journal_entry_uuid"`
	Body             *VoidJournalEntryJSONRequestBody
}

type VoidJournalEntryResponseObject interface {
	VisitVoidJournalEntryResponse(w http.ResponseWriter) error
}

type VoidJournalEntry200JSONResponse VoidJournalEntryResponse

func (response VoidJournalEntry200JSONResponse) VisitVoidJournalEntryResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(200)

	return json.NewEncoder(w).Encode(response)
}

type VoidJournalEntry400JSONResponse struct{ BadRequestJSONResponse }

func (response VoidJournalEntry400JSONResponse) VisitVoidJournalEntryResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(400)

	return json.NewEncoder(w).Encode(response)
}

type VoidJournalEntry404JSONResponse struct{ NotFoundJSONResponse }

func (response VoidJournalEntry404JSONResponse) VisitVoidJournalEntryResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(404)

	return json.NewEncoder(w).Encode(response)
}

type VoidJournalEntry500JSONResponse struct {
	InternalServerErrorJSONResponse
}

func (response VoidJournalEntry500JSONResponse) VisitVoidJournalEntryResponse(w http.ResponseWriter) error {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(500)

	return json.NewEncoder(w).Encode(response)
}

// StrictServerInterface represents all server handlers.
type StrictServerInterface interface {
	// Register a new fund source
	// (POST /v1/treasury/fund-sources)
	RegisterFundSource(ctx context.Context, request RegisterFundSourceRequestObject) (RegisterFundSourceResponseObject, error)
	// Record Journal Entries (the fund source's balance will be affected)
	// (POST /v1/treasury/fund-sources/{fund_source_uuid}/journal-entries)
	RecordJournalEntries(ctx context.Context, request RecordJournalEntriesRequestObject) (RecordJournalEntriesResponseObject, error)
	// Void a journal entry by creating a reverse journal entry
	// (POST /v1/treasury/fund-sources/{fund_source_uuid}/journal-entries/{journal_entry_uuid}/void)
	VoidJournalEntry(ctx context.Context, request VoidJournalEntryRequestObject) (VoidJournalEntryResponseObject, error)
}

type StrictHandlerFunc = strictecho.StrictEchoHandlerFunc
type StrictMiddlewareFunc = strictecho.StrictEchoMiddlewareFunc

func NewStrictHandler(ssi StrictServerInterface, middlewares []StrictMiddlewareFunc) ServerInterface {
	return &strictHandler{ssi: ssi, middlewares: middlewares}
}

type strictHandler struct {
	ssi         StrictServerInterface
	middlewares []StrictMiddlewareFunc
}

// RegisterFundSource operation middleware
func (sh *strictHandler) RegisterFundSource(ctx echo.Context) error {
	var request RegisterFundSourceRequestObject

	var body RegisterFundSourceJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.RegisterFundSource(ctx.Request().Context(), request.(RegisterFundSourceRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RegisterFundSource")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(RegisterFundSourceResponseObject); ok {
		return validResponse.VisitRegisterFundSourceResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// RecordJournalEntries operation middleware
func (sh *strictHandler) RecordJournalEntries(ctx echo.Context, fundSourceUuid FundSourceUUID) error {
	var request RecordJournalEntriesRequestObject

	request.FundSourceUuid = fundSourceUuid

	var body RecordJournalEntriesJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.RecordJournalEntries(ctx.Request().Context(), request.(RecordJournalEntriesRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "RecordJournalEntries")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(RecordJournalEntriesResponseObject); ok {
		return validResponse.VisitRecordJournalEntriesResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}

// VoidJournalEntry operation middleware
func (sh *strictHandler) VoidJournalEntry(ctx echo.Context, fundSourceUuid FundSourceUUID, journalEntryUuid JournalEntryUUID) error {
	var request VoidJournalEntryRequestObject

	request.FundSourceUuid = fundSourceUuid
	request.JournalEntryUuid = journalEntryUuid

	var body VoidJournalEntryJSONRequestBody
	if err := ctx.Bind(&body); err != nil {
		return err
	}
	request.Body = &body

	handler := func(ctx echo.Context, request interface{}) (interface{}, error) {
		return sh.ssi.VoidJournalEntry(ctx.Request().Context(), request.(VoidJournalEntryRequestObject))
	}
	for _, middleware := range sh.middlewares {
		handler = middleware(handler, "VoidJournalEntry")
	}

	response, err := handler(ctx, request)

	if err != nil {
		return err
	} else if validResponse, ok := response.(VoidJournalEntryResponseObject); ok {
		return validResponse.VisitVoidJournalEntryResponse(ctx.Response())
	} else if response != nil {
		return fmt.Errorf("unexpected response type: %T", response)
	}
	return nil
}